requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.15.0
python-telegram-bot==13.3
pytz==2021.1
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.23
Telethon==1.20
telethon-session-sqlalchemy==0.2.16
tornado==6.1
tzlocal==2.1
